+++ contracts/tokens/GnoOsTokenVaultEscrow.sol
@@ -29,7 +29,7 @@ contract GnoOsTokenVaultEscrow is OsTokenVaultEscrow {
     address osTokenConfig,
     address initialOwner,
     address _authenticator,
-    uint256 _liqThresholdPercent,
+    uint64 _liqThresholdPercent,
     uint256 _liqBonusPercent,
     address gnoToken
   )
